initial refactor: put components into directory
