+++ .bumpversion.cfg
@@ -7,4 +7,4 @@ serialize = {major}.{minor}.{patch}
 
 [bumpversion:file:pyproject.toml]
 
-[bumpversion:file:app.py]
+[bumpversion:file:components/static.py]
